+++ lib/ref_inspector/database.ex
@@ -34,7 +34,7 @@ defmodule RefInspector.Database do
     {:reply, state.ets_tid, state}
   end
 
-  def handle_cast(:reload, _state) do
+  def handle_cast(:reload, %State{ets_tid: old_ets_tid}) do
     state = %State{ets_tid: create_ets_table()}
 
     database_files = Config.database_files()
@@ -54,6 +54,8 @@ defmodule RefInspector.Database do
             |> store_refs(acc_state)
         end
       end)
+
+    :ok = drop_ets_table(old_ets_tid)
 
     {:noreply, state}
   end
@@ -84,6 +86,16 @@ defmodule RefInspector.Database do
     ets_opts = [:protected, :ordered_set, read_concurrency: true]
 
     :ets.new(ets_name, ets_opts)
+  end
+
+  defp drop_ets_table(ets_tid) do
+    true =
+      case :ets.info(ets_tid) do
+        :undefined -> true
+        _ -> :ets.delete(ets_tid)
+      end
+
+    :ok
   end
 
   defp store_refs([], state), do: state